共有関連はshareに画面を統一
